Minor cleanup for server stop CLI commands
- moved common stop functions to new cli/common/stop.py file
- renamed private function _validate_input to public and more verbose name since it's also used in other files now
- removed unused imports
