+++ vantage6/vantage6/cli/server/stop.py
@@ -5,7 +5,7 @@ from vantage6.common import info
 from vantage6.cli.common.decorator import click_insert_context
 from vantage6.cli.config import CliConfig
 from vantage6.cli.context.server import ServerContext
-from vantage6.cli.utils import helm_uninstall, stop_port_forward
+from vantage6.cli.common.stop import helm_uninstall, stop_port_forward
 
 
 @click.command()
